Pull item selection out of exam item service to remove circular dependencies
